Calculator
You can perform all basic operations (+, -, /, *, **, root). All errors are processed and displayed in the front. I also made an exception filter. The project has a connection to the database and migration. I thought about adding all the records to the database, but I didn’t have time to fully implement it